pyproject.toml
[tool.poetry]
name = "kino-torrent"
version = "0.1.0"
description = ""
authors = ["andimeon <andimeon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28.1"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
python-telegram-bot = "^21.10"
fastapi = {extras = ["standard"], version = "^0.115.6"}
ipython = "^8.31.0"
cryptography = "^44.0.0"
colorlog = "^6.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
